gitops code refactoring
